Example 'open subvi front panel/programmatically at run-time/': Implement that subVI front panels are opened in the monitor where the main.vi front panel is located
Problem: The monitor setup and the front panel window configuration used when closing the subVI determines where its front panel appears when it is called by main.vi. For example, if the front panel of a subvi was located at a secondary monitor located right of the primary monitor and is then saved and closed, the subVI front panel may not be visible when calling it from a setup where the secondary monitor is located left of the primary monitor.

Solution: Set "FP.Monitor"-property of subVI to value of the value of main.vi before calling the subVI.

Other possible approach (not used): Change "Window Run-Time Position"-VI Property (Ctrl + I) of subVI to "Position: Centered", "Monitor: Primary". However, while the front panel of subVI is always opened in a visible area, if the front panel of main.vi is not located on the primary monitor, you have to switch monitors to operate the subVI.